mini refactor move common method tests to child-class
